screenshot_verifier/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pillow>=10.0
numpy>=1.26
opencv-python-headless>=4.10
scikit-image>=0.22
torch>=2.0
timm>=0.9
albumentations>=1.4
pydantic>=2.5
python-multipart>=0.0.8
loguru>=0.7